- refactor(app): Batch 5 route buckets + RAF cleanup + prefetch sort immutability
- Move draft shared assets, app shell, nav/support trio, and character UI into
  _components/_lib; rewire imports and path-sensitive tests.
- Cancel requestAnimationFrame on unmount in CharacterDetails and
  MobileCharacterDetails progress animations.
- Sort a copy in usePriorityPrefetch to avoid mutating caller route arrays.
