+++ src/app/_components/characters/CharacterInfoCard.tsx
