+++ src/app/_components/characters/CharacterListOptimized.tsx
