+++ src/app/_components/characters/DesktopCharacterCard.tsx
