+++ src/app/_components/characters/InfoStatsCard.tsx
