+++ src/app/_components/characters/MobileCharacterTile.tsx
