+++ src/app/_components/characters/MobileCharacterTileSkeleton.tsx
