+++ src/app/_components/characters/RealmStatsCard.tsx
