+++ src/app/_components/characters/TotalStatsCard.tsx
