+++ src/app/draft/_lib/types.ts
@@ -1,4 +1,4 @@
-import { Id } from "../../../convex/_generated/dataModel";
+import { Id } from "../../../../convex/_generated/dataModel";
 
 export interface DraftPlayer {
   _id: Id<"draftPlayers">;
