+++ src/app/layout.tsx
@@ -2,12 +2,12 @@ import React from "react";
 import "./globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import Navbar from "./navbar";
+import Navbar from "./_components/navbar";
 import { ClerkProvider } from "@clerk/nextjs";
-import Footer from "./footer";
+import Footer from "./_components/footer";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
-import { Providers } from "./providers";
+import { Providers } from "./_components/providers";
 import { Toaster } from "sonner";
 import SupportPromptModal from "@/components/support/SupportPromptModal";
 import SignedOutNudge from "@/components/auth/SignedOutNudge";
