+++ src/app/user/[name]/characters/page.tsx
@@ -16,7 +16,7 @@ import {
 } from "@/server/publicUserCharacters";
 
 const CharacterListOptimized = dynamic(
-  () => import("@/app/components/CharacterListOptimized"),
+  () => import("@/app/_components/characters/CharacterListOptimized"),
   {
     loading: () => {
       const CharacterListSkeleton = require("@/app/user/_components/CharacterListSkeleton").default;
